Move a handful of other generally applicable components
