+++ ui/apps/platform/src/Components/CvssFormatted.tsx
