+++ ui/apps/platform/src/Components/DynamicIcon.tsx
