+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/Tables/AffectedDeploymentsTable.tsx
@@ -7,8 +7,8 @@ import { gql } from '@apollo/client';
 import useSet from 'hooks/useSet';
 import { UseURLSortResult } from 'hooks/useURLSort';
 import { VulnerabilityState } from 'types/cve.proto';
+import { DynamicColumnIcon } from 'Components/DynamicIcon';
 import { getWorkloadEntityPagePath } from '../../utils/searchUtils';
-import { DynamicColumnIcon } from '../../components/DynamicIcon';
 import EmptyTableResults from '../components/EmptyTableResults';
 import DeploymentComponentVulnerabilitiesTable, {
     DeploymentComponentVulnerability,
@@ -17,7 +17,7 @@ import DeploymentComponentVulnerabilitiesTable, {
     imageMetadataContextFragment,
 } from './DeploymentComponentVulnerabilitiesTable';
 import SeverityCountLabels from '../../components/SeverityCountLabels';
-import DateDistance from '../../components/DateDistance';
+import DateDistance from '../../../../Components/DateDistance';
 import { VulnerabilitySeverityLabel } from '../../types';
 
 export type DeploymentForCve = {
