+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/Tables/ImagesTable.tsx
@@ -6,12 +6,12 @@ import { Flex, Label } from '@patternfly/react-core';
 
 import { UseURLSortResult } from 'hooks/useURLSort';
 import { EyeIcon } from '@patternfly/react-icons';
+import { DynamicColumnIcon } from 'Components/DynamicIcon';
+import TooltipTh from 'Components/TooltipTh';
 import ImageNameLink from '../components/ImageNameLink';
 import SeverityCountLabels from '../../components/SeverityCountLabels';
-import { DynamicColumnIcon } from '../../components/DynamicIcon';
 import EmptyTableResults from '../components/EmptyTableResults';
-import DateDistance from '../../components/DateDistance';
-import TooltipTh from '../components/TooltipTh';
+import DateDistance from '../../../../Components/DateDistance';
 import { VulnerabilitySeverityLabel, WatchStatus } from '../../types';
 import ImageScanningErrorLabel from '../components/ImageScanningErrorLabelLayout';
